Exit non-zero to generate an "NA" testing report.
Since this module is only usable on threaded perls, it should not be processed
by CPANtesters runs where the perl is unthreaded.  Per discussion on
IRC #toolchain, the best way to generate an 'NA' result is to die() in the
Makefile.PL or, as here, Build.PL.

die() would cause "at line ..." to be included in the error output, so to
preserve the existing error output I'm suggesting simply exiting with a
non-zero value.